projeto/requirements.txt
Pillow
pandas
openpyxl
python-dotenv
streamlit
